requirements.txt
# Django and REST framework
Django>=5.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=24.1

# Swagger/OpenAPI documentation
drf-yasg>=1.21.7

# Database
psycopg2-binary>=2.9.9  # For PostgreSQL support
pymysql>=1.1.0  # For MySQL support (easier to install than mysqlclient on shared hosting)
cryptography>=41.0.0  # Required for MySQL 8.0+ authentication (caching_sha2_password)

# Date/Time handling
pytz>=2024.1

# Phone number validation
django-phonenumber-field>=7.3.0
phonenumbers>=8.13.30

# Environment variables
python-dotenv>=1.0.0

# Development tools
django-debug-toolbar>=4.3.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
factory-boy>=3.3.0

# Code formatting and linting
black>=24.1.1
flake8>=7.0.0
isort>=5.13.2

# API request handling
requests>=2.31.0

# Logging
python-json-logger>=2.0.7

# Scheduling
apscheduler>=3.10.0

# WSGI Server (for production)
gunicorn>=21.2.0
